Add the total time for all the tests if they were run serially
Summary: Our test runner runs quickly because we fork off processes up to the number of
CPU cores. This gives us //parallelism//.

It might be useful to know how long the tests took if you were to run them all
serially. Basically adding the times together you see if you give the `--verbose`
option to the runner.

But since state is not shared amongst the various process forks, we have to come
up with a mechanism to save the times and retrieve them later. So I used files.
Thought about memcache, but that might not be available to everyone, possibly.

{sync, type="parent", child="external", childrevid="34509", childdiffid="180717", parentdiffid="6742503"}

Reviewed By: @DmitrySoshnikov

Differential Revision: D1889961