Fix bug in ini_iterate
Summary: I wasn't considering actually passing a full setting to the function like

```
hhvm.ip_block_map.0.ip.deny.1
```

As it was, the dot parts would start with `hhvm` and then move to `ip_block_map`,
which would both be null and cause problems.

So instead, what we do is concatenate the dot parts starting from the beginning
until we get a valid pointer, then start iterating for the final value.

{sync, type="parent", child="external", childrevid="34809", childdiffid="181113", parentdiffid="6753118"}

Reviewed By: @fredemmott

Differential Revision: D1904982