PUNT in makeExit if instruction branches to itself
Summary: I recently made makeExit call makeExitSlow in these cases, but this
doesn't work while generating the initial region guards because
makeExitSlow needs the currentNormalizedInstruction (in HTS) which
isn't initialized at that time.  Instead, this diff uses the PUNT
mechanism to get the same effect of doing an InterpOne.

Also fix an issue with branchesToItself found during testing
(instrJumpOffset() would fail an assert if called on a Switch
instruction).

Reviewed By: @mhahnenberg

Differential Revision: D1904749