Block Access of Type Constants at Runtime
Summary: Currently the value of a type constant a string representation of its assigned
type. Since we are likely to change this representation I want to deny access
to type constants at runtime. To do this I thread set the flag that a constant
is a type constant that was introduced in {D1890946}.

Depends on D1890946

Reviewed By: @elgenie

Differential Revision: D1894885