Don't use exceptions to perform returns from interpreter to TC
Summary: This diff modifies the relevant interpreter functions to return a TCA
indicating when we should return to an address in the TC, rather than throwing
an exception. This adds a few instructions to the normal dispatchBB() path, but
greatly reduces the cost when the jitReturn logic actually kicks in.

This won't matter most of the time, but the unwinder does some locking and
under very heavy load this can cause contention and performance issues.

Reviewed By: @edwinsmith

Differential Revision: D1903700