Relax the rules relating to memory support somewhat in refcount-opts
Summary: I've tried to update the docs to describe how things work after this
change ignoring how they worked before this change, but I'll use the summary
here to describe the change relative to how things used to work.

One part of the change here fixes at least two existing bugs in how
`pessimized' flags worked, by just pessimizing everything any time we pessimize
anything (this was surprisingly in the noise in perflab).  The other (main)
change here is that prior to this diff, we had the invariant that
aset.lower_bound >= aset.memory_support.count() at all times for any aset.

This diff removes this rule, which mostly is an advantage for the "second case"
in the "Effects of Pure Stores on Memory Support" section, which now no longer
needs to set pessimize flags.  This comes up a lot when transfering values from
the eval stack to locals (e.g. after a function call), since this pass can't
generally see enough evidence that the eval stack slot is dead in time to keep
the old invariant.

A side-effect of this is that the merge function for memory_support has to be
able to handle more complex situations.  It previously relied on the
lower_bound always being high enough to reduce it enough to to only keep the
intersection of incoming memory support vectors, but this may not always be the
case now (the incoming edge could even already have a lower bound of zero now).
To handle this it now can do pessimization during merges in some situations,
but otherwise try to keep the old merge behavior when it still applies.

One other fallout is that a pure load from a location that's marked avail by
the mrinfo can't always be sure that it can remove_support from a previously
supported set.  If the previously supported set has a lower bound of zero now,
then we can't reduce its lower bound in order to remove the memory support bit,
which means we can't raise the lower bound in the new loaded set or we'd
potentially be violating the exclusivity rule.  We could pessimize the old set
in this situation, but there's a decent chance it may-alias this set anyway,
and we'd be worse off than we would be by doing nothing.  So this diff just
leaves state alone in that situation.

Reviewed By: @ottoni, @swtaarrs

Differential Revision: D1915041