Keep sweeping until both lists are empty
Summary: There are two sweeplists: for Sweepable and NativeData. The idea with
the sweep phase is to keep sweeping objects until the lists are empty,
becuase some sweep functions add or remove sweep entries. However,
we were processing each list in sequence. If a nativeData added a new
Sweepable, it wouldn't be swept.

This merges the sweeping so both lists are drained, and also moves
the Sweepable list into MemoryManager, so the logic is consolidated.

Reviewed By: @​bnell

Differential Revision: D1909125