ini_iterate should just return a valid value
Summary: `ini_iterate()` was focusing too much on strings. And while this will
be the end result of iterating over an ini map in most cases, sometimes,
we will just want the actual object.

Also, removed a bug where if the dot size was 0, we return nullptr. Removing
the `isString()` check above this handles this case anyway.

Reviewed By: @markw65

Differential Revision: D1912332