Make simplify() work as a subroutine.
Summary: Implement a function which replaces a post-gen instruction with its
simplifier result stream, in a way which guarantees consistency of
iterators.

The correctness of this replacement requires a number of additional
constraints on the simplifier.  In particular, the "output" of the
simplifier result stream must match the output of the original
instruction in the following ways:

- Any out edges on the last instruction of the stream must be reachable
  from the original instruction.  (The converse need not hold.)

- Only the last instruction of the stream may end a block, and then only
  if the original instruction ended a block.  (Again, the converse need
  not hold.)

All of these are conditions which already hold for all simplify
routines.  In addition, we need to ensure that the dst (unless it is a
src or a cns) is produced by one of the instructions in the output
stream, which we enforce by adding a Mov if necessary.

Reviewed By: @swtaarrs

Differential Revision: D1886363