Fix translateRegion for regions containing unnatural loops and unreachable paths
Summary: translateRegion was assuming that, if a region block is unreachable by
the time it tries to process it, then the block was definitely
unreachable.  The region blocks are sorted in a quasi-topological
order (rpo), which "almost" gives that guarantees, but not quite.  If
the region has an unnatural loop (ie, with multiple entries), one of
the loop entries may be unreachable when we first visit, but later
become reachable through another entry of that same loop.

To fix this issue, this diff changes translateRegion to keep a
work-queue of blocks to be processed instead of doing a single pass
through all the blocks.  If a block is not reachable when we try to
process it, then it's put back at the end of the queue for later
consideration.  We stop when there's no reachable block left in the
queue.

This diff also renames some things that were talking about "loop
headers" to say "blocks with unprocessed predecessors", which better
reflect what's going on.

Reviewed By: @jdelong

Differential Revision: D1906185