Add catch traces to PHP-level calls
Summary: This diff adds catch traces to all PHP-level calls (most notably
bindcall). This is necessary for temps spanning calls and variable spill space,
both of which will come in later diffs.

The C++ unwinder doesn't know how to reconstruct the value of %rsp while
unwinding through frames in the TC. We told it that the value is unchanged, but
rather than just leaving it along, it sets it to the CFA, which is somewhere on
the VM stack (or the heap, for resumables). To work around this, I added a mini
unwinder in unwind-x64.cpp that unwinds through TC frames, calling catch traces
as appropriate. This is also necessary to smash m_savedRip of unwound TC frames
at the right time - we were previously smashing them in tc_unwind_personality
but that was too early, causing us to incorrectly think the frame didn't have a
catch trace.

Reviewed By: @edwinsmith

Differential Revision: D1885174