Run class initializers before ObjectData constructor
Summary: During object construction we might need to run an initializer (like
86pinit) which can end up executing arbitrary PHP via autoload.  If we do this
inside the ObjectData constructor then we haven't initialized the object at this
point, so we have garbage in memory that could theoretically be read.  This
definitely crashes objprof, but it might also be an issue if we throw and try
to destroy the incomplete object.

Reviewed By: @alexmalyshev

Differential Revision: D1902574