Replace rpoSortCfgWithIds() with numberBlocks()
Summary: This seems to compose better, since it's usable with both poSortCfg
and rpoSortCfg, and bothers me a little less than redoing an rpo sort just
because something (usually findDominators) needs ids also.  Also make it not
require the BlockList to be stored with the ids, because you might already have
a block list (or only need to store the list and not the ids).

Reviewed By: @swtaarrs

Differential Revision: D1921928