Fix ExecutionContext::preventReturnsToTC() to play nice with bindcall
Summary: We now have a rule that if you enter a VM frame in the TC, you must
leave that frame either by returning to the caller in the TC, or by throwing an
exception, which gives the unwinder a chance to do any necessary
cleanup. preventReturnsToTC() as used by the debugger violates this rule by
blindly smashing m_savedRip in ActRecs to point at return helpers. This diff
changes it to instead use a new helper that runs the catch trace attached to
the old return address, then resumes in the interpreter. The unwinder knows to
look for this new helper and can find the catch trace to run when needed.

Reviewed By: @markw65

Differential Revision: D1910062