Get rid of BC marker passed through for FrameStateMgr::clearForUnprocessedPred
Summary: The BC marker being passed in one spot had a bad frame pointer, so it
was triggering an assert in clearForUnprocessedPred in wholecfg mode.
So I ended up getting rid of the marker altogether: it was just used
to reset the curFunc and spOffset, but these don't need to be cleared
in the presence of an unprocessed predecessor (they must agree in all
incoming paths).

Reviewed By: @alexmalyshev

Differential Revision: D1920727