Fixup token_get_all() output for altered `else if` tokens
Summary: https://github.com/facebook/hhvm/issues/2699
was "fixed" by altering lexer output to turn a sequence of
T_ELSE T_WHITESPACE T_IF into T_ELSEIF.

This was simple and elegant, but it causes the output of
token_get_all() to be... slightly unexpected.

Unwind the rewrite in token_get_all() so that everything looks
as it should.

Closes #4796

Reviewed By: @DmitrySoshnikov

Differential Revision: D1921219