Fix static-initialization-fiasco as shown in issue #4954. Also correct a comment.
Summary: libp11-kit.so.0 also uses thread specific keys and because it
calls pthread_key_create() very very early, it gets "assigned" a key
with value 0.
Next when the ICU shared library loads, it calls setlocale() during its
init. setlocale uses the stack-allocated object g_thread_safe_locale_handler
which doesn't have any pointer inside, so it calls ThreadLocal::create()
to create a new ThreadSafeLocaleHandler object. But
ThreadLocalManager::s_manager isn't fully initialized yet, so s_manager's
m_key is 0 and when s_manager::setTop() is called, g_thread_safe_locale_handler
is set into the key. This is the same key as is used by libp11-kit.
As the rest of the HHVM binary loads, ThreadLocalManager::s_manager is properly
initialized and a new ThreadSafeLocaleHandler is created and everything works
fine.
When the process starts to shutdown and p11-kit is unloaded,
pthread_getspecific() with key 0 is called here and it now has the pointer to
the stack object g_thread_safe_locale_handler. When free() is called on this,
there's a SEGV.