fix simplifyCallBuiltin() operating on collections
Summary: this started when i tried to implement the same optimization for $collection->count() as count($collection). this seemed trivial since there already was an isEmpty() optimization. except it didn't work because the conditions were never met.

the 0th arg to CallBuiltin is FramePtr, the 2nd argument is where the collection could possibly be.

Reviewed By: @elgenie

Differential Revision: D1930392