Immediately close socket if there's no final hook
Summary: Previously, the channel would get closed by the hook, which
only runs when there are inotify events which re-trigger parsing. With a
normal build, this will basically always happen, and so is not a big
deal. However, if there's some error case or something else where the
build didn't run, it may be a while until we re-run parsing, leaving the
`hh_client` hanging.

Instead, if there's no hook, just directly close the channel. The
assumption is that if we have a hook, we've done stuff that will make
sure it runs (otherwise why bother with the hook).

This matters for an upcoming diff which will immediately kill the build
if usernames mismatch, and without this fix we'll leave `hh_client`
hanging in that error case.

{sync, type="parent", child="external", childrevid="35535", childdiffid="185229", parentdiffid="6871462"}

Reviewed By: @paulcavallaro

Differential Revision: D1930781