Disable ext_string hasExactlyOneRef opts
Summary: There's not a great way to do this sort of thing a from a builtin
function.  If you do it while taking a const String& (what the code apparently
originally did), it potentially breaks for C++ callers, because they will leave
out increfs and it may modify strings in unexpected ways.  OTOH, taking String
by value is dangerous for the JIT (and it caused a bug too: D1515837), because
the JIT doesn't run the real ~String (no reload of m_px or null check), and
String-by-value lets you call many member functions that could change the m_px.
So if these optimization matter, we should re-do them in a way that only
affects the JIT (e.g. special optimized runtime helpers called from the JIT
only).

Reviewed By: @swtaarrs

Differential Revision: D1931958