Initalize the memory manager in RuntimeOption::Load
Summary: Some settings and initializations require it now, particularly the
`HardwareCounter::Init()`

If we don't have it, then the C++ tests will fail with an assert.

Don't need it in hhhbc now, since it was put there before the call
to `RuntimeOption::Load()`, but we have it directly in there now.

Reviewed By: @markw65

Differential Revision: D1933313