In ob_get_status() fix type key
Summary: As documented and implemented in PHP PHP, type=0 is internal, and type=1 is for user handlers. This function got it the wrong way around.

This is an untested fix for https://phabricator.wikimedia.org/T89918 . MediaWiki relies on seeing type=1 to know which output buffers it can safely cancel in wfResetOutputBuffers(). When HHVM gives type=0 instead, MediaWiki fails to disable output buffering, and thus hits an OOM when it tries to stream out large output, e.g. a 200MB PDF file in the current test case.
Closes https://github.com/facebook/hhvm/pull/4868

Reviewed By: @JoelMarcey

Differential Revision: D1862059