Don't attempt a Live translation after failing an Optimize translation
Summary: This adds extra, unnecessary complexity, and the live state may not
correspond to the start state specified in the RegionDesc (due to
guarded types).  Plus, this code path seems to be involved in
mysterious crashes I sometimes see in perflab with wholecfg mode.
Instead, this diff just bails to the interpreter in such cases.

Reviewed By: @swtaarrs

Differential Revision: D1931016