Strengthen checkCfg for checking SSA-ness of locally-defined names
Summary: It checked that all sources of an instruction are defined in a block
that dominate the instruction, but if that block happened to be the same block
as the use, it didn't check if the def came earlier than the use.

Note: I don't think anything checks for multiple definitions right now.

Reviewed By: @edwinsmith

Differential Revision: D1936053