allow undefined data-* and aria-* XHP attributes with ->:
Summary: This is documented behavior, and allowed by the XHP library.

This currently only works in hh_client because of the 'assume every XHP client has a PHP superclass' hack.

If you open :ui:base in the IDE, it brings this up as an error, as it isn't aware of the .hhconfig

{sync, type="parent", child="external", childrevid="35757", childdiffid="186897", parentdiffid="6933436"}

Reviewed By: @elgenie

Differential Revision: D1936815