More efficient functions for reading/writing via afdt sockets
Summary: I've got a couple of diffs that need to do reads and
writes over an afdt socket, so I decided to pull out some
utility functions. While doing that I realized we could use
the scatter/gather mechanism of sendmsg/recvmsg to make
the transfers more efficient, and write wrappers to make
it easy to send/receive multiple items in one go.

This also fixes a bug in recv_fd, where it passed an
uninitialized length into afdt_recv_fd_msg. I think we
got away with it because send_fd is always followed either
by a read, a poll or another send_fd; the result is that
there's never any data to fetch.

Reviewed By: @jdelong

Differential Revision: D1933219