Make BCMarker members private
Summary: It's in this weird hybrid state with both accessors and public
members; choose one.  (Public might make more sense, but for now this is
fine---I think I need to give this more information about inlined frames, so it
might help to have the abstraction layer.  If not we can remove it later.  Also
seems like the dummy thing sorta gets in the way of having it as a pure POD...)

Reviewed By: @mhahnenberg

Differential Revision: D1939743