Fix several issues with AKExists
Summary: For an object base, it can raise warnings and needs a catch trace, and
also should not be dce'able.  The key arg was "Cell" but it actually only
support Int, Null and Str subtypes.  Split the opcodes into two cases so they
can have different flags.  And when the array key is a constant string that is
known not to be an integer, skip that check dynamically, and if it is known to
be strictly an integer, call the integer version.

Reviewed By: @swtaarrs

Differential Revision: D1942704