Move context from WaitableWaitHandle into a union
Summary: Since D1936935 moved m_blockable to the subclasses, we have got some space in
the union that can be used when the WaitHandle is not finished yet. Context
index is a good candidate for that and this single byte is currently
occupying 8 bytes as part of WaitableWaitHandle. Let's move it to the union and
reverse the expected memory regression of D1936935.

Reviewed By: @jdelong

Differential Revision: D1937645