Update BC marker before emitting type-checks for a region block
Summary: Otherwise, we may end up trying to create an invalid marker because
the current functions hasn't been updated.  Note that we were setting
the offset for the exit block properly but, without this diff, we
could have the wrong Func in the marker in case we didn't generate IR
for the return of an inlined callee (which happens if it turns out to
be unreachable).

Reviewed By: @jdelong

Differential Revision: D1942148