Fix subtyping of Tunresolveds in contravariant positions
Summary: In general, if a supertype is `Tunresolved`, then we consider it to
be "not yet finalized", but if a subtype is `Tunresolved` and the
supertype isn't, we've probably hit a type annotation and should
consider the supertype to be definitive.

However, sometimes we want this behavior reversed, e.g. when the type
annotation has a contravariant generic parameter or a `super` constraint --
now the definitive type is the subtype.

Reviewed By: @elgenie, @jwatzman

Differential Revision: D1948639