Don't redefine StkPtrs for HHIR call instructions
Summary: Removes the StkPtr destinations from IR call instructions.  To do this
without regressing things, this diff removes almost all cross-trace uses of
rVmSp for non-resumed functions.  In a non-resumed function, we now don't
expect rbx to have any particular value on entry to a region, and don't
synchronize it on most exits from a region---the remaining exception is in the
return helper unique stubs (which means this diff still requires every region
to synchronize rbx before doing a vret).  We also still synchronize it in some
places where it is acting kinda like an argument register (before interp ones
and in FCallArray).

The scheme in this diff is this: DefSP in a non-generator does an lea from
rVmFp into a Vreg, and in a resumed function it just generates a defsp (to
receive a live rVmSp arg on entry to the region).  The IR still has a StkPtr
from DefSP, but doesn't redefine it in call instructions.  In a resumed
function, the Vreg for the StkPtr will be spilled across calls.  In a
non-resumed function, vasm-copy will rewrite references to the DefSP Vreg to be
relative to rbp instead.

Remaining work:

   - We need to stop redefining StkPtrs for inlined functions, so we can stop
     TRACE_PUNTing wholecfg translations with changing stkptrs at merge points.

   - We should stop synchronizing rVmSp before returns.

   - rStashedAR is very apparently unnecessary now---we should probably be able
     to just pass the callee rbp to prologues.

   - (Maybe not soon) resumed functions could use a "stack base" register that
     doesn't need to be adjusted at region boundaries.

Reviewed By: @swtaarrs

Differential Revision: D1999090