finalize Closure methods in the IDL
Summary: user-supplied code is not supposed to be extending Closure.

one way to trick the runtime into allowing "class X extends Closure" is to override the constructor to prevent the parent (`c_Closure::t___construct()`) from erroring out.

this creates two ways of crashing:
- by overriding the constructor, a Closure object can be created without the `CreateCl` bytecode being executed, thus c_Closure::init() is never run. `c_Closure->m_func` is `nullptr`. `var_dump($closure_obj)` enters `c_Closure::t__debuginfo()` and null is dereferenced. stack trace from such a crash P19780626.
- hhbbc expects that classes that extend Closure have one and only one method named __invoke. stack traces from this crash P19780671 and P19780675

this diff does not directly address the hhbbc assertions.

Reviewed By: @markw65

Differential Revision: D1949712