Change GenericRetDecRefs helpers to act like normal C++ functions
Summary: This makes them callable from LLVM without creating another custom
calling convention, and greatly simplifies codegen. The new version prevents an
add/sub of %rsp around calls to the stub, slightly decreasing code size in the
main translation. But now the stub has to push a few things on the native stack
whenever we actually destroy a value (which is about 9% of values we decref in
the profiling I did).

Reviewed By: @edwinsmith

Differential Revision: D1956474