Don't use CollectionType::Invalid in ArrayIter::InitFuncPtr table.
Summary: Explicitly convert the invalid case to a known index, then use
a new ObjectData::collectionType() accessor that avoids a redundant
check for non-collections. New code is equivalent to old, but doesn't
depend on the CollectionType::Invalid enum.

Adds a slimmed down ObjectData::collectionType() accessor that
asserts isCollection(). Followon diffs will use it when we already
know we have a collection.

Reviewed By: @jano

Differential Revision: D1956656