Convert PackedArray capacity to a 16-bit pseudo-floating-point format.
Summary: Instead of a 24-bit value with two encodings (small and big), use
a 16-bit coding with an 11-bit mantissa and 5-bit exponent, so the
decode is (m<<e). This keeps the property that the encoded pattern
of small values is the same as the decoded pattern, but saves one
header byte.

Reviewed By: @markw65

Differential Revision: D1973541