Improve behavior of hack lambdas for uninit vars
Summary: Lambda auto-capturing has a bug with capturing vars which are still in uninit state.

A basic example:

```
$foo = () ==> { // warning here for $x
  $x = 2;
  var_dump($x);
};

$x = 1;
```

Another examples can be found in the test, one of which an exception parameter of a `catch` block.

The diff adds a new `CUGetL` ("Cell or Uninit Get from Local"), that don't issue a warning for uninit vars, and don't promote them to `null`. The var though is still captured in lambdas as uninit to issue a warning on read of this var in the closure (but not on write, see test examples).

Reviewed By: @jdelong

Differential Revision: D1913667