Cache result of \HH\Client\typecheck
Summary: Take advantage of the stamp file added by D1960376 in order to
cache the `\HH\Client\typecheck` function. This dramatically reduces the
average time to run it from about 100ms down to 0.1ms or so when no code
is changing. This is fast enough that the next step is going to be to
automatically run that function when invoking Hack code (and, if you're
as big as FB and that speed actually does matter, you can just turn the
checking off).

I also note a few areas for potential future improvement and even better
perf -- and also one problem: due to wanting to keep this code stupid
simple, it doesn't deal quite right with multiple subdirectories with
`.hhconfig` files in them. My plan is to deal with that if people
complain, and hopefully get to a point where you don't need an
`hhconfig` in the default case anyways so it becomes more moot.

Reviewed By: @sgolemon, @paulbiss

Differential Revision: D1960438