Distinguish between strings and arrays and add new stats
Logger calls now distinguish between arrays and strings so we have
more information to deal with. All stats are now given as a total,
and relating specifically to only arrays or strings.

Changed the 1-bit stat to inlcude the number of copies under refcounting
for clarity.

Formatted output in a table.