Fix FrameState's refineLocalType and refineStackType
Summary: When these methods actually refine the known type, we need to drop the
old value.  Otherwise, we may end up using a value with a more general
type than what has been checked/asserted for that location.

This bug seems to be around for a long time, but we generally got
around it because, when a value is known for a local/stack, the
preoptimize functions for CheckLoc/CheckStk/AssertLoc/AssertStk would
turn these instructions into CheckType/AssertType, and then update()
would call refineLocalValue/refineStackValue when processing these new
instructions, which would then fix the tracked values.
However, this doesn't work when we're using FrameState in fixed-point
mode for the reoptimize pass, because the effect of a
CheckLoc/CheckStk that is turned into a CheckType by the reoptimize
pass won't update the value of the corresponding local/stack slot for
its successor block.

Reviewed By: @jdelong

Differential Revision: D1950781