Remove defaulted arguments from BackEnd virtuals
Summary: Mixing virtuals and defaulted arguments is a little too subtle, since
the derived classes don't really inherit the defaults (you have to keep them
all matching or it'll depend on the static type when you invoke the member
function).  Convert to private virtual idiom to separate interface and
implementation.

Reviewed By: @edwinsmith

Differential Revision: D1956347