Collapse vasm return instructions
Summary: Having separate pushm{} and ret{} vasm instructions during the return
sequence creates a brief window where spill space must be freed and it's
illegal to spill/fill. This was awkward but ok in D1910725, but only because we
never generate vasm that needs to fill right before a ret{}. That's going to
change, so this diff modifies the return-related hhir/vasm instructions to get
rid of the gap between pushm and ret.

RetCtrl now does everything that the LdRetAddr/FreeActRec/RetCtrl sequence used
to to, and is used for returning in most cases. It lowers to the new vretm{}
instruction in codegen, which atomically (from vasm's point of view) pushes the
return address and returns. This makes it safe for vretm{} to use a src that
was spilled.

Resumed async functions need to DecRef the resumable containing the current
ActRec, so they still need the LdRetAddr/FreeActRec/DecRef sequence in order to
not read from a destroyed object. LdRetAddr has been changed to a real load,
allowing the push/ret sequence to still be atomic, this time in the
AsyncRetCtrl hhir instruction and vret{} vasm instructions. This does add an
extra instruction to the non-eager async function return path (it's now "load;
...; push; ret" rather than "pushm; ...; ret") but hopefully that won't matter
in practice.

Reviewed By: @jdelong

Differential Revision: D1952523