Get rid of Eval.HHIRBytecodeControlFlow runtime option
Summary: D1823069 had already unified translateRegion's behavior that was
dependent on this runtime option, and this option was mostly left as a
(hacky) way to add some non-linear control flow on top of the hottrace
region selector.  But that approach generates more limited control
flow than wholecfg does these days, so we can kill that now.  There
were just a few places left checking for Eval.HHIRBytecodeControlFlow,
which this diff replaces with checks for the region selector and
Eval.JitLoops.

I'm still leaving the hacky way that we add loop backedges on top of
hottrace when Eval.JitLoops is on.  This is just to increase test
coverage for now when loops are enabled.

Reviewed By: @bertmaher

Differential Revision: D1956768