Fix vasm-print to not have a catch-all case for imm()
Summary: Several things were calling overloads that probably weren't intended
(e.g. it looked like "unsigned bit" from tbcc (whatever that is) was printing
?), and several types were missing.  Immed and Immed64 were causing ambiguous
resolution (leading to calling the catchall) in unintended cases because they
convert from a bunch of integral types (not the nicest thing to mix that with
overloading that is trying to catch some of those types, but we can get away
with enable_if).  While here, bindjcc was missing an I(trflags).

Reviewed By: @swtaarrs

Differential Revision: D1959935