Move emitTraceCall out of all the BackEnds, use an IR instruction instead
Summary: Remove emitTraceCall proliferation.  They had a BackEnd virtual but
for no apparent reason, as all callsites have known callees (but not really
because the arm BackEnd isn't final).  Also it apparently was using
X64Assembler still.

That would be reason enough to change, but more importantly: after we dereserve
rbx, it needs to use information we have during irgen to figure out how to
materialize a top-of-stack pointer to make the call, so it needs to come from
that level instead of being automatically added before we generate vasm.

Reviewed By: @swtaarrs

Differential Revision: D1961847