DOM: Add check doc guard to normalizeDocument
Summary: Fixes #4893

We have the `CHECK_DOC` (and its counterpart `CHECK_WRITE_DOC`) guard which is put in every method/getter that accesses some state data on a document. This was forgotten in the `normalizeDocument` method.

(Unfortunatelly, usage of `CHECK_DOC` is hardcoded with the `obj` variable name instead of passing it as a parameter, so I introduced `CHECK_THIS_DOC` that directly accepts `this_`, and `CHECK_DOC` calls it)

{sync, type="parent", child="external", childrevid="35235", childdiffid="183477", parentdiffid="6829873"}

Reviewed By: @paulbiss

Differential Revision: D1921353