Use health monitor for pagelets
Summary: We want to not start running too many pagelet threads if the server is
low on memory.  We already have HostHealthMonitor on for normal requests, so
this just applies it to pagelets.

Most of this diff is just doing the work to move HostHealthMonitor from fb-only
to open source.  The only real change aside from code movement is to make it so
the health monitor doesn't have to know all the metric types up front; they're
now declared as statics and add themselves to a singleton vector, and the
health monitor can pick them up from there.

Reviewed By: @markw65

Differential Revision: D1955840