Fix mapping the shared area of RDS in Cygwin.
Summary: MapViewOfFileEx() requires "the specified memory region is not already in
use by the calling process" when mapping the shared area below. Otherwise
it will return MAP_FAILED. We first map the full size to make sure the
memory area is available. Then we unmap and map the lower portion of the
RDS at the same address.

Part of #5068.
Closes https://github.com/facebook/hhvm/pull/5069

Reviewed By: @paulbiss

Differential Revision: D1937899