Add timestamps to logs
Summary: This makes it easier to read logs

Also have the logger provide a `measure` function that records how long
some function takes; this factors out a bunch of repeated logic

Note that there are now bits of code in the init / recheck sequences
that aren't included in any our phase timings; this is mostly because
the parsing-naming-typecheck sequence passes around a lot of state that
is difficult to fully wrap in a closure argument to `Hh_logger.measure`.
However, the things we are measuring should comprise the bulk of the
time taken. Also this has forced me to clean things up a little :)

Reviewed By: @jwatzman

Differential Revision: D1967131