Don't use rVmSp for stack overflow checks
Summary: The maxStackCells() already includes locals and iterators, so we can
just compare using the rStashedAR.  (When nPassed > nparams, we'll be shuffling
some of them out into an extra-args, and the caller already checked that we
didn't overflow for the nPassed).  This should actually prevent us from
overflowing in some situations we didn't really need to, but otherwise should
be the same.

Reviewed By: @edwinsmith

Differential Revision: D1973492