Support "super" constraints
Summary: This diff adds support for "super" constraints on typevars. "super X" works
similarly to "as X", except that it constrains a typevar to be a supertype
of X (as opposed to constraining it to be a subtype of X).

One of the main use cases for "super" comes up when dealing with generic
classes/interfaces that are covariant on one or more type parameters. Here
is a simple example:

  interface Iterable<+T> {
    public function concat<Tu super T>(Iterable<Tu> $x): Iterable<Tu>;
  }

In the example above, Iterable is covariant on its type parameter T. This
means that "T" or "Iterable<T>" cannot be used as a parameter type for any
of Iterable's methods (i.e. "function concat(Iterable<T> $x): Iterable<T>"
is not allowed). By introducing a new typevar "Tu" and constraining it to
be a supertype of T, the programmer is able to write certain kinds of APIs
that otherwise would not be possible.

The main motivation for this feature is to make it possible to remove all
of the collection covariance hacks and to make collections use Hack's
formal covariance/contravariance features (+/-).

Reviewed By: @jwatzman

Differential Revision: D1903420