Fix variance checks for type constraints
Summary: An `as` constraint is an upper type bound. For a given Foo<T>,
suppose `Tu as B` is satisfied. Then if B is a subtype of A,
`Tu as A` is also satisfied. So Foo is contravariant w.r.t. T.

Reversing the above tells us that `super` constraints should be
covariant.

Reviewed By: @jwatzman, @elgenie, @dlreeves

Differential Revision: D1948638