Check that variance of tparams in parent classes are compatible
Summary: As one might expect, a child class can have an invariant type param if
the parent has the same type param marked as co/contravariant, but not
vice versa.

While at it, I factored out the conversion of `hint` -> `ty`,
so that Typing_variance would not need to return a modified env.

Reviewed By: @jwatzman

Differential Revision: D1972296