Don't use rVmSp in emitMagicFuncPrologue
Summary: This does an additional load and some more ALU ops, but it's an
unusual case (and we're about to call C++ to do heavy lifting).  Some of this
code sequence could probably be moved into a magic-call-specific version of the
MakePacked helper, but it appears not to move the needle in perflab so this is
probably fine.

Reviewed By: @edwinsmith

Differential Revision: D1976379